Move onButtonClick to Item
