+++ src/ItemGrid/ItemGrid.js
@@ -6,19 +6,13 @@ import 'react-grid-layout/css/styles.css';
 import 'react-resizable/css/styles.css';
 
 import './ItemGrid.css';
-import PluginItem from '../Item/PluginItem/Item';
+import { Item } from '../Item/Item';
 
 import { gridRowHeight, getGridColumns, gridCompactType } from './gridUtil';
-import {
-    getPluginByType,
-    getFavoriteObjectFromItem,
-    getPluginItemConfig,
-    onPluginItemResize,
-} from './pluginUtil';
+import { onPluginItemResize } from './pluginUtil';
 
 import { orObject } from '../util';
 import * as fromReducers from '../reducers';
-import { apiFetchFavorite } from '../api';
 import ModalLoadingMask from '../widgets/ModalLoadingMask';
 
 const { fromSelected } = fromReducers;
@@ -41,19 +35,13 @@ export class ItemGrid extends Component {
     };
 
     render() {
-        const {
-            isLoading,
-            dashboardItems,
-            onButtonClick,
-            onItemResize,
-            edit,
-        } = this.props;
+        const { isLoading, dashboardItems, onItemResize, edit } = this.props;
 
         if (!dashboardItems.length) {
             return <div style={{ padding: 50 }}>No items</div>;
         }
 
-        const pluginItems = dashboardItems.map((item, index) => {
+        const items = dashboardItems.map((item, index) => {
             const expandedItem = this.state.expandedItems[item.id];
             let hProp = { h: item.h };
 
@@ -77,7 +65,7 @@ export class ItemGrid extends Component {
                         onItemResize(newItem.i);
                     }}
                     className="layout"
-                    layout={pluginItems}
+                    layout={items}
                     cols={getGridColumns()}
                     rowHeight={gridRowHeight}
                     width={window.innerWidth}
@@ -85,22 +73,17 @@ export class ItemGrid extends Component {
                     isDraggable={edit}
                     isResizable={edit}
                 >
-                    {pluginItems
-                        .filter(item => getFavoriteObjectFromItem(item)) //TODO IMPROVE
-                        .map(item => {
-                            return (
-                                <div key={item.i} className={item.type}>
-                                    <PluginItem
-                                        item={item}
-                                        editMode={edit}
-                                        onButtonClick={onButtonClick}
-                                        onToggleItemFooter={
-                                            this.onToggleItemFooter
-                                        }
-                                    />
-                                </div>
-                            );
-                        })}
+                    {items.map(item => {
+                        return (
+                            <div key={item.i} className={item.type}>
+                                <Item
+                                    item={item}
+                                    editMode={edit}
+                                    onToggleItemFooter={this.onToggleItemFooter}
+                                />
+                            </div>
+                        );
+                    })}
                 </ReactGridLayout>
             </div>
         );
@@ -118,18 +101,6 @@ ItemGrid.defaultProps = {
 // Container
 
 const currentItemTypeMap = {}; //TODO: improve
-
-const onButtonClick = (id, type, targetType) => {
-    const plugin = getPluginByType(targetType);
-
-    apiFetchFavorite(id, type).then(favorite => {
-        const itemConfig = getPluginItemConfig(favorite, true);
-
-        plugin.load(itemConfig);
-
-        currentItemTypeMap[id] = targetType;
-    });
-};
 
 const onItemResize = id => {
     if (
@@ -151,7 +122,6 @@ const mapStateToProps = state => {
         dashboardItems,
         isLoading: sGetSelectedIsLoading(state),
         edit: sGetSelectedEdit(state),
-        onButtonClick,
         onItemResize,
     };
 };
